ریفکتور کد جهت اجرای پروژه Ordering بدون مشکل اجرایی

+++ src/Ordering/Ordering-Core/DTO/OrderItemDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace Ordering.Components.DTO;
 
-public record OrderItemDto(Guid ProductId, string ProductName, int Quantity, decimal Price);
+public record OrderItemDto(Guid ProductId, string ProductName, int Quantity, int Price);
